shrinking things down to 3x5+2
oof. this is going to be tough to learn methinks.

the inspiration for the enter combo is from ASETNIOP where all 4 fingers
on the right hand is enter. that's a lot of keys with heavy switches,
but it's a start at least, and I can always change things around if I
want to later.