only use middle and ring for enter
while the 4 finger combo works fine, with heavy switches it's just a tad
much. so move down to 2 finger combo. I'm hoping the combo trigger time
is fast enough that I won't trigger it too many times using rolls on nt
which is a suuuuuuper common roll for me, but we shall see. experiments!